Fixed mirror list declaration for the non-fiction book route.
Fixed an issue that caused the book mirror list to fail its declaration due to the non-book mirror list being declared globally, changed to the correct mirror_list name: "for mirror in mirror_list". This fixes an issue leading to an infinite loop being called as the libgen webscraper tries to find the element on the incorrect page.