Fix UnboundLocalError of the on_reaction_add event listener data list
The on_reaction_add event listener encountered an UnboundLocalError when the condition 'user.id in search_responses' was not met, resulting in the 'data' variable being referenced before assignment. To resolve this issue, a default value for 'data' has been provided before the if statement, ensuring that the variable is defined in all code paths. This at the very least will prevent an UnboundLocalError from occurring.